new beta_counts
